- fixed some bugs in techinasia
