+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "jarviscrawlercore",
-  "version": "0.2.127",
+  "version": "0.2.128",
   "description": "jarvis crawler core",
   "main": "index.js",
   "scripts": {
